refactor: divide PriceCalculator into asking and calculating parts
- add MembershipAsker to deal with asking part of PriceCalculator
- update the package names of service to show their jobs better
