+++ src/main/java/store/Application.java
@@ -3,14 +3,15 @@ package store;
 import store.controller.StoreController;
 import store.domain.Products;
 import store.domain.Promotions;
-import store.service.PriceCalculator;
 import store.service.PromotionTimer;
-import store.service.ordermanager.FreeAdditionAsker;
-import store.service.ordermanager.NonPromotionAsker;
-import store.service.stockmanager.FreeProductsChecker;
-import store.service.stockmanager.OrdersChecker;
-import store.service.stockmanager.StockChecker;
-import store.service.stockmanager.StockDeductor;
+import store.service.StockDeductor;
+import store.service.asker.FreeAdditionAsker;
+import store.service.asker.MembershipAsker;
+import store.service.asker.NonPromotionAsker;
+import store.service.dtocreator.FreeProductsChecker;
+import store.service.dtocreator.OrdersChecker;
+import store.service.dtocreator.PriceCalculator;
+import store.service.dtocreator.StockChecker;
 import store.util.Retrier;
 import store.util.md.MdReader;
 import store.view.InputView;
@@ -26,9 +27,11 @@ public class Application {
         Promotions promotions = new Promotions();
         Retrier retrier = new Retrier();
         PromotionTimer timer = new PromotionTimer(products, promotions);
-        PriceCalculator calculator = new PriceCalculator(inputView, products, promotions, timer, retrier);
+
         FreeAdditionAsker freeAdditionAsker = new FreeAdditionAsker(inputView, products, promotions, timer, retrier);
         NonPromotionAsker nonPromotionAsker = new NonPromotionAsker(inputView, products, promotions, timer, retrier);
+        MembershipAsker membershipAsker = new MembershipAsker(inputView, retrier);
+        PriceCalculator calculator = new PriceCalculator(products, promotions, timer, membershipAsker);
         StockChecker stockChecker = new StockChecker(products);
         OrdersChecker ordersChecker = new OrdersChecker(products);
         FreeProductsChecker freeProductsChecker = new FreeProductsChecker(products, promotions, timer);
