+++ src/main/java/store/controller/StoreController.java
@@ -6,14 +6,14 @@ import store.domain.Products;
 import store.domain.Promotions;
 import store.domain.input.Orders;
 import store.dto.receipt.ReceiptFreeProductDTO;
-import store.service.PriceCalculator;
 import store.service.PromotionTimer;
-import store.service.ordermanager.FreeAdditionAsker;
-import store.service.ordermanager.NonPromotionAsker;
-import store.service.stockmanager.FreeProductsChecker;
-import store.service.stockmanager.OrdersChecker;
-import store.service.stockmanager.StockChecker;
-import store.service.stockmanager.StockDeductor;
+import store.service.StockDeductor;
+import store.service.asker.FreeAdditionAsker;
+import store.service.asker.NonPromotionAsker;
+import store.service.dtocreator.FreeProductsChecker;
+import store.service.dtocreator.OrdersChecker;
+import store.service.dtocreator.PriceCalculator;
+import store.service.dtocreator.StockChecker;
 import store.util.Retrier;
 import store.util.md.MdKeywords;
 import store.util.md.MdReader;
