+++ src/test/java/store/service/dtocreator/FreeProductsCheckerTest.java
@@ -1,4 +1,4 @@
-package store.service.stockmanager;
+package store.service.dtocreator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
